refactor(test): split test/retry.ts into smaller files
