+++ test/retry/delay.ts
@@ -1,0 +1,574 @@
+import {setTimeout as delay} from 'node:timers/promises';
+import test from 'ava';
+import ky from '../../source/index.js';
+import {createHttpTestServer} from '../helpers/create-http-test-server.js';
+import {withPerformance} from '../helpers/with-performance.js';
+
+const fixture = 'fixture';
+const defaultRetryCount = 2;
+const retryAfterOn500 = 2;
+const retryAfterOn413 = 2;
+
+test('respect Retry-After: 0 and retry immediately', async t => {
+	const retryCount = 4;
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.writeHead(413, {
+				'Retry-After': 0,
+			});
+
+			response.end('');
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 4 + 4 + 4 + 4,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: retryCount,
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+});
+
+test('RateLimit-Reset is treated the same as Retry-After', async t => {
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === defaultRetryCount + 1) {
+			response.end(fixture);
+		} else {
+			const header = (requestCount < 2) ? 'RateLimit-Reset' : 'Retry-After';
+			response.writeHead(429, {
+				[header]: 1,
+			});
+
+			response.end('');
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 1000 + 1000,
+		async test() {
+			t.is(await ky(server.url).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 3);
+});
+
+test('RateLimit-Reset with time since epoch', async t => {
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === defaultRetryCount + 1) {
+			response.end(fixture);
+		} else {
+			const twoSecondsByDelta = 2;
+			const oneSecondByEpoch = (Date.now() / 1000) + 1;
+			response.writeHead(429, {
+				'RateLimit-Reset': (requestCount < 2) ? twoSecondsByDelta : oneSecondByEpoch,
+			});
+
+			response.end('');
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 2000 + 1000,
+		async test() {
+			t.is(await ky(server.url).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 3);
+});
+
+test('respect 413 Retry-After', async t => {
+	const startTime = Date.now();
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === defaultRetryCount + 1) {
+			response.end((Date.now() - startTime).toString());
+		} else {
+			response.writeHead(413, {
+				'Retry-After': retryAfterOn413,
+			});
+			response.end('');
+		}
+	});
+
+	const timeElapsedInMs = Number(await ky(server.url).text());
+	t.true(timeElapsedInMs >= retryAfterOn413 * 1000);
+	t.is(requestCount, retryAfterOn413 + 1);
+});
+
+test('respect 413 Retry-After with timestamp', async t => {
+	const startTime = Date.now();
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t, {bodyParser: false});
+	server.get('/', (_request, response) => {
+		requestCount++;
+		if (requestCount === defaultRetryCount + 1) {
+			response.end((Date.now() - startTime).toString());
+		} else {
+			// @NOTE we need to round up to the next second due to http-date resolution
+			const date = new Date(Date.now() + ((retryAfterOn413 + 1) * 1000)).toUTCString();
+			response.writeHead(413, {
+				'Retry-After': date,
+			});
+			response.end('');
+		}
+	});
+
+	const timeElapsedInMs = Number(await ky(server.url).text());
+	t.true(timeElapsedInMs >= retryAfterOn413 * 1000);
+	t.is(requestCount, retryAfterOn413 + 1);
+});
+
+test('doesn\'t retry on 413 without Retry-After header', async t => {
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+		response.sendStatus(413);
+	});
+
+	await t.throwsAsync(ky(server.url).text(), {message: /Payload Too Large/});
+	t.is(requestCount, 1);
+	await ky(server.url, {throwHttpErrors: false}).text();
+	t.is(requestCount, 2);
+});
+
+test('respect custom `afterStatusCodes` (500) with Retry-After header', async t => {
+	const startTime = Date.now();
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === defaultRetryCount + 1) {
+			response.end((Date.now() - startTime).toString());
+		} else {
+			response.writeHead(500, {
+				'Retry-After': retryAfterOn500,
+			});
+			response.end('');
+		}
+	});
+
+	const timeElapsedInMs = Number(await ky(server.url, {retry: {afterStatusCodes: [500]}}).text());
+	t.true(timeElapsedInMs >= retryAfterOn500 * 1000);
+	t.is(requestCount, retryAfterOn500 + 1);
+});
+
+test('respect maxRetryAfter', async t => {
+	const retryCount = 4;
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.writeHead(413, {
+				'Retry-After': 1,
+			});
+
+			response.end('');
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 420 + 420 + 420 + 420,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: {
+					limit: retryCount,
+					maxRetryAfter: 420,
+				},
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+
+	requestCount = 0;
+
+	await withPerformance({
+		t,
+		expectedDuration: 1000 + 1000 + 1000 + 1000,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: {
+					limit: retryCount,
+					maxRetryAfter: 2000,
+				},
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+});
+
+test('invalid Retry-After header falls back to retry delay', async t => {
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+		if (requestCount === 1) {
+			response.writeHead(429, {
+				'Retry-After': 'not-a-valid-value',
+			});
+			response.end('');
+			return;
+		}
+
+		response.end(fixture);
+	});
+
+	const result = await Promise.race([
+		ky(server.url, {
+			timeout: false,
+			retry: {
+				limit: 1,
+				delay: () => 10,
+			},
+		}).text(),
+		delay(500).then(() => {
+			throw new Error('Retry took too long');
+		}),
+	]);
+
+	t.is(result, fixture);
+	t.is(requestCount, 2);
+});
+
+test('respect maximum backoffLimit', async t => {
+	const retryCount = 4;
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 300 + 600 + 1200 + 2400,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: retryCount,
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+
+	requestCount = 0;
+
+	await withPerformance({
+		t,
+		expectedDuration: 300 + 600 + 1000 + 1000,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: {
+					limit: retryCount,
+					backoffLimit: 1000,
+				},
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+});
+
+test('backoffLimit: undefined treats as no limit (Infinity)', async t => {
+	const retryCount = 4;
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	// When backoffLimit is undefined, it should behave the same as no limit
+	// (i.e., delays should not be clamped, same as default behavior)
+	await withPerformance({
+		t,
+		expectedDuration: 300 + 600 + 1200 + 2400,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: {
+					limit: retryCount,
+					backoffLimit: undefined,
+				},
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+});
+
+test('respect custom retry.delay', async t => {
+	const retryCount = 4;
+	let requestCount = 0;
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await withPerformance({
+		t,
+		expectedDuration: 200 + 300 + 400 + 500,
+		async test() {
+			t.is(await ky(server.url, {
+				retry: {
+					limit: retryCount,
+					delay: n => 100 * (n + 1),
+				},
+			}).text(), fixture);
+		},
+	});
+
+	t.is(requestCount, 5);
+});
+
+test('jitter: true applies full jitter to delay', async t => {
+	const retryCount = 3;
+	let requestCount = 0;
+	const delays: number[] = [];
+	let lastTime = Date.now();
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		const now = Date.now();
+		if (requestCount > 0) {
+			delays.push(now - lastTime);
+		}
+
+		lastTime = now;
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await ky(server.url, {
+		retry: {
+			limit: retryCount,
+			jitter: true,
+		},
+	}).text();
+
+	t.is(requestCount, 4);
+
+	// Full jitter should produce delays between 0 and the computed delay
+	// Add 50% tolerance for system overhead and CI variability
+	t.true(delays[0] >= 0 && delays[0] <= 450);
+	t.true(delays[1] >= 0 && delays[1] <= 900);
+	t.true(delays[2] >= 0 && delays[2] <= 1800);
+});
+
+test('jitter: custom function applies custom jitter', async t => {
+	const retryCount = 3;
+	let requestCount = 0;
+	const jitterCalls: number[] = [];
+
+	const customJitter = (delay: number) => {
+		jitterCalls.push(delay);
+		return delay * 0.5; // Half the delay
+	};
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await ky(server.url, {
+		retry: {
+			limit: retryCount,
+			jitter: customJitter,
+		},
+	}).text();
+
+	t.is(requestCount, 4);
+	t.is(jitterCalls.length, 3); // Called for each retry
+
+	// Verify the jitter function received the correct delays
+	t.is(jitterCalls[0], 300); // First retry
+	t.is(jitterCalls[1], 600); // Second retry
+	t.is(jitterCalls[2], 1200); // Third retry
+});
+
+test('jitter respects backoffLimit', async t => {
+	const retryCount = 3;
+	let requestCount = 0;
+	const delays: number[] = [];
+	let lastTime = Date.now();
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		const now = Date.now();
+		if (requestCount > 0) {
+			delays.push(now - lastTime);
+		}
+
+		lastTime = now;
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await ky(server.url, {
+		retry: {
+			limit: retryCount,
+			backoffLimit: 500,
+			jitter: true,
+		},
+	}).text();
+
+	t.is(requestCount, 4);
+
+	// With backoffLimit of 500, all delays should be <= 500ms
+	// Even though the computed delays would be 300, 600, 1200
+	// After jitter and backoffLimit, they should all be <= 500
+	// Add 50% tolerance for system overhead and CI variability
+	t.true(delays[0] >= 0 && delays[0] <= 750);
+	t.true(delays[1] >= 0 && delays[1] <= 750);
+	t.true(delays[2] >= 0 && delays[2] <= 750);
+});
+
+test('jitter works with custom delay function', async t => {
+	const retryCount = 2;
+	let requestCount = 0;
+	const jitterCalls: number[] = [];
+
+	const customJitter = (delay: number) => {
+		jitterCalls.push(delay);
+		return delay * 0.5;
+	};
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === retryCount + 1) {
+			response.end(fixture);
+		} else {
+			response.sendStatus(500);
+		}
+	});
+
+	await ky(server.url, {
+		retry: {
+			limit: retryCount,
+			delay: n => 100 * n, // Custom delay: 100ms, 200ms, etc.
+			jitter: customJitter,
+		},
+	}).text();
+
+	t.is(requestCount, 3);
+	t.is(jitterCalls.length, 2);
+	t.is(jitterCalls[0], 100); // First retry with custom delay
+	t.is(jitterCalls[1], 200); // Second retry with custom delay
+});
+
+test('jitter is not applied when Retry-After header is present', async t => {
+	const startTime = Date.now();
+	let requestCount = 0;
+	const jitterCalls: number[] = [];
+
+	const customJitter = (delay: number) => {
+		jitterCalls.push(delay);
+		return delay * 0.5;
+	};
+
+	const server = await createHttpTestServer(t);
+	server.get('/', (_request, response) => {
+		requestCount++;
+
+		if (requestCount === 3) {
+			response.end((Date.now() - startTime).toString());
+		} else {
+			response.writeHead(429, {
+				'Retry-After': 1,
+			});
+			response.end('');
+		}
+	});
+
+	const timeElapsedInMs = Number(await ky(server.url, {
+		retry: {
+			jitter: customJitter,
+		},
+	}).text());
+
+	// Should have made 3 requests (initial + 2 retries)
+	t.is(requestCount, 3);
+	// Jitter function should NOT have been called when Retry-After is present
+	t.is(jitterCalls.length, 0);
+	// Should have waited at least 2 seconds (1s per retry)
+	t.true(timeElapsedInMs >= 2000);
+});
